add date parsing and logging
Added the momentJs library in order to allow for easier date-handling on the client-side.  We'll want to come up w/ a date strategy for the real version, but for the prototype, this works well.

Also moved the console.log to a LogHelper class in order for us to have more control over tracing functionality (i.e. logging to a server, turning off, etc...)